refactor(team): split team-routes into lib/routes/{agents,teams,context}
Pure relocation of populateTeamRoutes (32 routes, 1,049 lines) into three
populate functions plus a shared response-schema module, following the
workflows lib/routes/* template. Route bodies are byte-identical; only
import depths and the populate wiring changed.

Registration order is load-bearing: the RouteRegistry breaks
specificity-score ties by insertion order, so agents register first,
then context, then org-team routes — keeping every tie pair
(/context/files, /context/team, /teams/team, ...) resolving exactly as
before. Populate functions (not workflows-style exported const arrays)
because the agent-create route needs the indexAgentStatic dep injected
from index.ts, and a const array would force a circular import.
